🎨(project) split the app monolith in separated modules
We should be able to use streamers and models in another context than
the web application.
